+++ src/components/CalendarView.tsx
@@ -1,0 +1,164 @@
+import {
+  Stack,
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TableHead,
+  TableRow,
+  Typography,
+} from '@mui/material';
+
+import { EventItem } from './EventItem';
+import { CalendarViewType, Event } from '../types';
+import {
+  formatDate,
+  formatMonth,
+  formatWeek,
+  getEventsForDay,
+  getWeekDates,
+  getWeeksAtMonth,
+} from '../utils/dateUtils';
+
+const weekDays = ['일', '월', '화', '수', '목', '금', '토'];
+
+interface CalendarViewProps {
+  view: CalendarViewType;
+  currentDate: Date;
+  filteredEvents: Event[];
+  notifiedEvents: string[];
+  holidays: Record<string, string>;
+}
+
+export const CalendarView = ({
+  view,
+  currentDate,
+  filteredEvents,
+  notifiedEvents,
+  holidays,
+}: CalendarViewProps) => {
+  const renderWeekView = () => {
+    const weekDates = getWeekDates(currentDate);
+    return (
+      <Stack data-testid="week-view" spacing={4} sx={{ width: '100%' }}>
+        <Typography variant="h5">{formatWeek(currentDate)}</Typography>
+        <TableContainer>
+          <Table sx={{ tableLayout: 'fixed', width: '100%' }}>
+            <TableHead>
+              <TableRow>
+                {weekDays.map((day) => (
+                  <TableCell key={day} sx={{ width: '14.28%', padding: 1, textAlign: 'center' }}>
+                    {day}
+                  </TableCell>
+                ))}
+              </TableRow>
+            </TableHead>
+            <TableBody>
+              <TableRow>
+                {weekDates.map((date) => (
+                  <TableCell
+                    key={date.toISOString()}
+                    sx={{
+                      height: '120px',
+                      verticalAlign: 'top',
+                      width: '14.28%',
+                      padding: 1,
+                      border: '1px solid #e0e0e0',
+                      overflow: 'hidden',
+                    }}
+                  >
+                    <Typography variant="body2" fontWeight="bold">
+                      {date.getDate()}
+                    </Typography>
+                    {filteredEvents
+                      .filter(
+                        (event) => new Date(event.date).toDateString() === date.toDateString()
+                      )
+                      .map((event) => {
+                        const isNotified = notifiedEvents.includes(event.id);
+                        return <EventItem key={event.id} event={event} isNotified={isNotified} />;
+                      })}
+                  </TableCell>
+                ))}
+              </TableRow>
+            </TableBody>
+          </Table>
+        </TableContainer>
+      </Stack>
+    );
+  };
+
+  const renderMonthView = () => {
+    const weeks = getWeeksAtMonth(currentDate);
+
+    return (
+      <Stack data-testid="month-view" spacing={4} sx={{ width: '100%' }}>
+        <Typography variant="h5">{formatMonth(currentDate)}</Typography>
+        <TableContainer>
+          <Table sx={{ tableLayout: 'fixed', width: '100%' }}>
+            <TableHead>
+              <TableRow>
+                {weekDays.map((day) => (
+                  <TableCell key={day} sx={{ width: '14.28%', padding: 1, textAlign: 'center' }}>
+                    {day}
+                  </TableCell>
+                ))}
+              </TableRow>
+            </TableHead>
+            <TableBody>
+              {weeks.map((week, weekIndex) => (
+                <TableRow key={weekIndex}>
+                  {week.map((day, dayIndex) => {
+                    const dateString = day ? formatDate(currentDate, day) : '';
+                    const holiday = holidays[dateString];
+
+                    return (
+                      <TableCell
+                        key={dayIndex}
+                        sx={{
+                          height: '120px',
+                          verticalAlign: 'top',
+                          width: '14.28%',
+                          padding: 1,
+                          border: '1px solid #e0e0e0',
+                          overflow: 'hidden',
+                          position: 'relative',
+                        }}
+                      >
+                        {day && (
+                          <>
+                            <Typography variant="body2" fontWeight="bold">
+                              {day}
+                            </Typography>
+                            {holiday && (
+                              <Typography variant="body2" color="error">
+                                {holiday}
+                              </Typography>
+                            )}
+                            {getEventsForDay(filteredEvents, day).map((event) => {
+                              const isNotified = notifiedEvents.includes(event.id);
+                              return (
+                                <EventItem key={event.id} event={event} isNotified={isNotified} />
+                              );
+                            })}
+                          </>
+                        )}
+                      </TableCell>
+                    );
+                  })}
+                </TableRow>
+              ))}
+            </TableBody>
+          </Table>
+        </TableContainer>
+      </Stack>
+    );
+  };
+
+  return (
+    <>
+      {view === CalendarViewType.WEEK && renderWeekView()}
+      {view === CalendarViewType.MONTH && renderMonthView()}
+    </>
+  );
+};
